requirements.txt
# Requirements needed for running scripts
robotframework >= 3.2.2
robotframework-pythonlibcore >= 3.0.0
robotframework-seleniumlibrary >= 3.3.1
pyotp >= 2.6.0
